Return in testing

+++ src/app/services/apiServices.ts
@@ -5,6 +5,12 @@ const SUMUP_API_KEY = process.env.SUMUP_API_KEY
 const SUMUP_MERCHANT_CODE = process.env.SUMUP_MERCHANT_CODE
 
 export async function createReaderCheckout (readerId: string, totalAmount: number, returnUrl: string): Promise<string | undefined> {
+	logger.silly('Creating reader checkout')
+
+	if (process.env.NODE_ENV === 'test') {
+		return 'test'
+	}
+
 	try {
 		const response = await axios.post(`https://api.sumup.com/v0.1/merchants/${SUMUP_MERCHANT_CODE}/readers/${readerId}/checkout`, {
 			total_amount: {
@@ -38,6 +44,12 @@ interface IPairReader {
 }
 
 export async function pairReader (pairingCode: string): Promise<IPairReader | undefined> {
+	logger.silly('Pairing reader')
+
+	if (process.env.NODE_ENV === 'test') {
+		return 'test' as any
+	}
+
 	try {
 		const response = await axios.post(`https://api.sumup.com/v0.1/merchants/${SUMUP_MERCHANT_CODE}/readers`, {
 			pairing_code: pairingCode
@@ -54,6 +66,12 @@ export async function pairReader (pairingCode: string): Promise<IPairReader | un
 }
 
 export async function unpairReader (readerId: string): Promise<boolean> {
+	logger.silly('Unpairing reader')
+
+	if (process.env.NODE_ENV === 'test') {
+		return true
+	}
+
 	try {
 		await axios.delete(`https://api.sumup.com/v0.1/merchants/${SUMUP_MERCHANT_CODE}/readers/${readerId}`, {
 			headers: {
